Upgrade to Java 21, install Gradle Wrapper, fix compilation errors
- Bump Java target from 17 to 21 (sourceCompatibility, targetCompatibility, jvmTarget)
- Upgrade AGP 8.2.0 → 8.3.2 for Java 21 jlink compatibility
- Install Gradle Wrapper (gradlew, gradlew.bat, gradle-wrapper.jar)
- Add proxy config in gradle.properties (127.0.0.1:10808)
- Add material-icons-extended dependency
- Fix 14 Kotlin compilation errors across 6 source files
- Add adaptive launcher icon resources (mipmap-anydpi-v26 + vector drawables)
- Configure Android SDK path in local.properties

+++ app/src/main/java/com/tiaosheng/counter/util/CalorieCalculator.kt
@@ -8,7 +8,7 @@ package com.tiaosheng.counter.util
 object CalorieCalculator {
 
     /** 跳绳 MET 值（中等速度 100-120 次/分钟） */
-    const val JUMP_ROPE_MET = 12.0
+    const val JUMP_ROPE_MET = 12.0f
 
     /**
      * @param totalJumps 有效跳跃次数
@@ -19,7 +19,7 @@ object CalorieCalculator {
     fun calculate(totalJumps: Int, weightKg: Float, durationSeconds: Int): Float {
         if (durationSeconds <= 0 || weightKg <= 0) return 0f
         val durationHours = durationSeconds / 3600f
-        return (JUMP_ROPE_MET * weightKg * durationHours).toFloat()
+        return JUMP_ROPE_MET * weightKg * durationHours
     }
 
     /**
